refactor: 예약 삭제/조회 컨트롤러 합치기
